refactor: 클래스명이 동사로 시작하지 안도록 수정
