--- src/main/java/lotto/domain/lotto/CreateLottoNumberStrategy.java
@@ -1,6 +1,0 @@
-package lotto.domain.lotto;
-
-@FunctionalInterface
-public interface CreateLottoNumberStrategy {
-    LottoNumber create();
-}
