--- src/main/java/lotto/domain/lotto/CreateLottoNumbersStrategy.java
@@ -1,8 +1,0 @@
-package lotto.domain.lotto;
-
-import java.util.List;
-
-@FunctionalInterface
-public interface CreateLottoNumbersStrategy {
-    List<LottoNumber> create(LottoNumber bonusNumber);
-}
