+++ src/main/java/lotto/domain/lotto/strategy/LottoNumberCreateStrategy.java
@@ -1,0 +1,8 @@
+package lotto.domain.lotto.strategy;
+
+import lotto.domain.lotto.LottoNumber;
+
+@FunctionalInterface
+public interface LottoNumberCreateStrategy {
+    LottoNumber create();
+}
